refactor(editor): serialization import pipeline 분리
